moved custom hooks in hooks directory, updated imports, added Link to Sale button on hero baner
